23課題の修正 ①フォーマットのずれを修正
