+++ src/main/java/raisetech/student/Repository/StudentRepository.java
@@ -1,0 +1,15 @@
+package raisetech.student.management;
+
+import java.util.List;
+import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Select;
+
+@Mapper
+public interface StudentRepository {
+
+	@Select("SELECT * FROM students")
+	List<Student> search();
+
+	@Select("SELECT * FROM student_courses")
+	List<StudentCourse> searchCourseList();
+}
